deleted index page, renamed guided.js
